refactor: 운동 신청 메서드 참여 서비스로 이동
